Bug 1927706 - Part 5: Update in-tree ICU to 76. r=platform-i18n-reviewers,gregtatum
Differential Revision: https://phabricator.services.mozilla.com/D227185

UltraBlame original commit: b8c177cfaeb24c9f65c70074563b1b209f8eb4bd

+++ intl/icu/source/common/ucnvlat1.cpp
@@ -381,7 +381,7 @@ ucnv_Latin1FromUTF8(UConverterFromUnicodeArgs *pFromUArgs,
             b=*source++;
             if(U8_IS_SINGLE(b)) {
                 
-                *target++=(uint8_t)b;
+                *target++ = b;
                 --targetCapacity;
             } else if( 
                        b>=0xc2 && b<=0xc3 &&
